Set e2e feature flags reliably
The manual calls to `beforeEach` were not guaranteerd to run before
other `beforeEach` calls - some of which did the initial page
navigation. This meant that some tests would sometimes be
interacting with pages that would have been loaded without setting
the x-forced-feature-flags header, and thus not have the feature
flags applied.

This method was sourced from
https://github.com/microsoft/playwright/issues/9468#issuecomment-943707670
