+++ src/e2e/fixtures/basePage.ts
@@ -15,6 +15,7 @@ import { DataBrokersPage } from "../pages/dataBrokersPage.js";
 import { AutomaticRemovePage } from "../pages/automaticRemovePage.js";
 
 const test = base.extend<{
+  sharedBeforeEach: void;
   landingPage: LandingPage;
   authPage: AuthPage;
   dashboardPage: DashboardPage;
@@ -26,6 +27,25 @@ const test = base.extend<{
   dataBrokersPage: DataBrokersPage;
   automaticRemovePage: AutomaticRemovePage;
 }>({
+  // See https://github.com/microsoft/playwright/issues/9468#issuecomment-943707670
+  // Sets the `x-forced-feature-flags` on every request to Monitor.
+  sharedBeforeEach: [
+    async ({ context }, use) => {
+      await context.route("**/*", async (route) => {
+        const requestUrl = route.request().url();
+        const headers = route.request().headers();
+
+        if (new URL(requestUrl).origin === process.env.E2E_TEST_BASE_URL) {
+          headers["x-forced-feature-flags"] =
+            "SidebarNavigationRedesign,EditScanProfileDetails";
+        }
+
+        await route.continue({ headers });
+      });
+      await use();
+    },
+    { scope: "test", auto: true },
+  ],
   // The `use` function is no React hook and not linted correctly.
   // For more info see issue: https://github.com/facebook/react/issues/31237
   /* eslint-disable react-hooks/rules-of-hooks */
