+++ src/e2e/specs/dashboard/dashboard-breaches-scan.spec.ts
@@ -4,7 +4,6 @@
 
 import { test, expect } from "../../fixtures/basePage.js";
 import { removeUnicodeChars } from "../../utils/helpers.js";
-import "../../utils/setFeatureFlags";
 
 // bypass login
 test.use({ storageState: "./e2e/storageState.json" });
