+++ src/e2e/specs/dashboard/dashboard-payment.spec.ts
@@ -4,7 +4,6 @@
 
 import { test } from "../../fixtures/basePage.js";
 import { checkAuthState } from "../../utils/helpers.js";
-import "../../utils/setFeatureFlags";
 
 // bypass login
 test.use({ storageState: "./e2e/storageState.json" });
