+++ src/e2e/specs/landing/landing-content.spec.ts
@@ -7,7 +7,6 @@ import {
   defaultScreenshotOpts,
   emailInputShouldExist,
 } from "../../utils/helpers.js";
-import "../../utils/setFeatureFlags";
 
 test.describe(`${process.env.E2E_TEST_ENV} - Verify the Landing Page content`, () => {
   test.beforeEach(async ({ landingPage }) => {
